refactor: ScaleBar into maps-compose-widgets package
