+++ README.md
@@ -199,19 +199,19 @@ and don't use `Marker` composables (unless you don't care about `onClick`
 events). Clustering is the only use-case tested with `MapEffect`, there may be
 gotchas depending on what features you use in the utility library.
 
-## Components
+## Widgets
 
-This library also provides composable components that you can use alongside the `GoogleMap` composable.
+This library also provides optional composable widgets in the `maps-compose-widgets` library that you can use alongside the `GoogleMap` composable.
 
 ### ScaleBar
 
-This composable shows the current scale of the map in feet and meters when zoomed into the map, changing to miles and kilometers, respectively, when zooming out. A `DisappearingScaleBar` is also included, which appears when the zoom level of the map changes, and then disappears after a configurable timeout period.
+This widget shows the current scale of the map in feet and meters when zoomed into the map, changing to miles and kilometers, respectively, when zooming out. A `DisappearingScaleBar` is also included, which appears when the zoom level of the map changes, and then disappears after a configurable timeout period.
 
 The [ScaleBarActivity](app/src/main/java/com/google/maps/android/compose/ScaleBarActivity.kt) demonstrates both of these, with the `DisappearingScaleBar` in the upper left corner and the normal base `ScaleBar` in the upper right:
 
 ![maps-compose-scale-bar-cropped](https://user-images.githubusercontent.com/928045/175665891-a0635004-2201-4392-83b3-0c6553b96926.gif)
 
-Both components leverage the `CameraPositionState` in maps-compose and therefore are very simple to configure with their defaults:
+Both versions of this widget leverage the `CameraPositionState` in `maps-compose` and therefore are very simple to configure with their defaults:
 
 ```kotlin
 ScaleBar(
@@ -250,7 +250,10 @@ dependencies {
     implementation 'com.google.android.gms:play-services-maps:18.0.2'
     
     // Also include Compose version `1.2.0-alpha03` or higher - for example:
-    implementation "androidx.compose.foundation:foundation:2.4.0-alpha03"
+    implementation 'androidx.compose.foundation:foundation:2.4.0-alpha03'
+    
+    // Optionally, you can include the widgets library if you want to use ScaleBar, etc.
+    implementation 'com.google.maps.android:maps-compose-widgets:1.0.0'
 }
 ```
