+++ app/build.gradle
@@ -59,6 +59,8 @@ dependencies {
     // declaration if you want to test the sample app with a Maven Central release of the library.
     //implementation  "com.google.maps.android:maps-compose:2.2.1"
     implementation project(':maps-compose')
+    //implementation  "com.google.maps.android:maps-compose-widgets:1.0.0"
+    implementation project(':maps-compose-widgets')
     implementation 'com.google.android.gms:play-services-maps:18.0.2'
 }
