+++ maps-compose-widgets/build.gradle
@@ -36,9 +36,12 @@ android {
 
 dependencies {
     implementation "androidx.compose.foundation:foundation:$compose_version"
+    implementation 'com.google.maps.android:maps-compose:2.2.1'
+    implementation 'androidx.compose.material:material:1.1.1'
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'com.google.android.gms:play-services-maps:18.0.2'
     implementation 'com.google.maps.android:maps-ktx:3.4.0'
+    implementation 'com.google.maps.android:maps-utils-ktx:3.3.0'
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
