+++ settings.gradle
@@ -9,3 +9,4 @@ dependencyResolutionManagement {
 rootProject.name = "android-maps-compose"
 include ':app'
 include ':maps-compose'
+include ':maps-compose-widgets'
